change primary data to snowflake

+++ database/migrations/2023_10_23_190936_create_players_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('players', function (Blueprint $table) {
             $table->id();
-            $table->integer('table_id');
+            $table->bigInteger('table_id');
             $table->integer('role_id')->nullable();
             $table->string('name');
             $table->integer('status')->default(1);
